move project files to namespaced path for sbt-in-sbt / IDE use; resolve conflicts
