+++ project/src/main/scala/mesosphere/maven/MavenSettings.scala
